Renamed vault-specific lib for more clarity.
