+++ src/lib/vault/index.js
